Complete TODO for splitting monolith of media.test.ts into their specific test files.
- Clean up imports to use the index versus the media file.
