--- src/media/media.ts
@@ -1,11 +1,0 @@
-export * from './processors/InstagramMediaProcessor';
-export * from './processors/InstagramImageProcessor';
-export * from './processors/InstagramVideoProcessor';
-export * from './processors/DefaultMediaProcessorFactory';
-export * from './interfaces/ProcessStrategy';
-export * from './interfaces/MIMEType';
-export * from './interfaces/InstagramPostProcessingStrategy';
-export * from './interfaces/ImageMediaProcessingStrategy';
-export * from './interfaces/VideoMediaProcessingStrategy';
-export * from './interfaces/MediaProcessorFactory';
-export * from './utils';
